Add custom hook in post
